requirements.txt
langchain
openai
chromadb
tiktoken
unstructured
bs4
selenium
Pillow
pytesseract
PyPDF2
libmagic
python-magic
frontend
tools
PyMuPDF
fitz